requirements.txt
openmeteo_requests===1.3.0
requests_cache===1.2.1
pandas===2.2.3
retry_requests===2.0.0
pygame===2.6.0
PyQt5===5.15.11
geocoder===1.38.1
flask===3.1.0
flask-cors===5.0.0
requests===2.32.3
python-dotenv===1.0.1